Change naming of Rust benches
`secure_cell_wrapped -> rust_secure_cell`
`Wrapped Secure Cell -> RustThemis - Secure Cell`
