+++ benches/themis/benches/rust_secure_cell_seal_passphrase.rs
@@ -14,32 +14,27 @@
 
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
 
-use themis::{keys::SymmetricKey, secure_cell::SecureCell};
+use themis::secure_cell::SecureCell;
 
-const CONTEXT: &[u8] = b"Themis Core benchmark";
+const CONTEXT: &[u8] = b"Rust Themis benchmark";
+const PASSPHRASE: &str = "correct horse battery staple";
 
 const KB: usize = 1024;
 const MB: usize = 1024 * KB;
 #[allow(clippy::identity_op)]
 const MESSAGE_SIZES: &[usize] = &[
-    16,     // UUID
-    32,     // SymmetricKey
+    // There is no significant difference between those due to KDF
     64,     // cache line (and close to EcdsaPrivateKey)
-    256,    // RsaPrivateKey
     4 * KB, // memory page
-    16 * KB,
-    64 * KB,
     1 * MB, // L2 cache
-    2 * MB,
-    4 * MB,
 ];
 
 pub fn encryption(c: &mut Criterion) {
-    let cell = SecureCell::with_key(SymmetricKey::new())
+    let cell = SecureCell::with_passphrase(PASSPHRASE)
         .expect("invalid key")
-        .token_protect();
+        .seal();
 
-    let mut group = c.benchmark_group("Wrapped Secure Cell encryption - Token Protect, master key");
+    let mut group = c.benchmark_group("RustThemis - Secure Cell encryption - Seal, passphrase");
     for message_size in MESSAGE_SIZES {
         group.throughput(Throughput::Bytes(*message_size as u64));
         group.bench_with_input(
@@ -58,22 +53,22 @@ pub fn encryption(c: &mut Criterion) {
 }
 
 pub fn decryption(c: &mut Criterion) {
-    let cell = SecureCell::with_key(SymmetricKey::new())
+    let cell = SecureCell::with_key(PASSPHRASE)
         .expect("invalid key")
-        .token_protect();
+        .seal();
 
-    let mut group = c.benchmark_group("Wrapped Secure Cell decryption - Token Protect, master key");
+    let mut group = c.benchmark_group("RustThemis - Secure Cell decryption - Seal, passphrase");
     for message_size in MESSAGE_SIZES {
         group.throughput(Throughput::Bytes(*message_size as u64));
         group.bench_with_input(
             BenchmarkId::from_parameter(pretty(*message_size)),
             message_size,
             |b, &size| {
-                let (encrypted, token) = cell
+                let encrypted = cell
                     .encrypt_with_context(vec![0; size], CONTEXT)
                     .expect("failed encryption");
                 b.iter(|| {
-                    let decrypted = cell.decrypt_with_context(&encrypted, &token, CONTEXT);
+                    let decrypted = cell.decrypt_with_context(&encrypted, CONTEXT);
                     assert!(decrypted.is_ok());
                 })
             },
